backend/requirements.txt
Django==3.2.3
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pyshorteners==1.0.1
Pillow==9.0.0
PyYAML==6.0
reportlab==4.4.0